move printDomToImage back to frontend
